Reduced huge bugs in logic

+++ headers/order.h
@@ -7,6 +7,6 @@ struct Order {
 	std::string id;
 	SIDE side;
 	unsigned int quantity;
-    double price;
+    float price;
 };
